June 5 by Yuchao Zhang

+++ src/main/java/com/example/restingspace/service/RoomService.java
@@ -20,8 +20,8 @@ public class RoomService {
         return roomDao.getRoom(rid);
     }
 
-    public void deleteRoom(long roomId) {
-        roomDao.deleteRoom((int) roomId);
+    public void deleteRoom(int roomId) {
+        roomDao.deleteRoom(roomId);
     }
 
     public void addRoom(Room room){
